Add temporary fix for stopping test cases
The problem is we can't cancel a timer from a callback of a different
timer (due to DPDK infra limitations). This is exactly what we've been
doing and was causing undefined behavior. This commit is just a dirty
hack and should never make it in dev/common! It's just a quick fix to
have the CI testing up and running. Once CI is ready we need to fix this
properly before merging.